Create data directory with sdist.
We store a copy of the schema in lib/rhev/data, to be shipped with
source distributions of python-rhev so that it can be used at run time.
That directoy is an empty directory, so it can't be added to git. This
change creates the directory as part of the "genschema" command.

Reported-by: Michael Pasternak <mpastern@redhat.com>