restapi: Mapping problem with _x64 OsTypes
When the user wants to update the Os-Type of a VM, he enters, for example: windows_2008x64 (the value as he saw it in 'capabilities'). Behind the scenes, we look for the matching OsType enum entry. The user's string undergoes toUppercase() and then matched against the enum's values. So in this example, we would have 'WINDOWS_2008X64' - large X matched against 'WINDOWS_2008x64' - small 'x'. This is not identified as a match, so the mapping silently fails.

Acked By: Ewoud Kohl van Wijngaarden <ewoud+rhevm-api@kohlvanwijngaarden.nl>